+++ CHANGELOG.rst
@@ -3,12 +3,14 @@ CHANGELOG
 
 7.0.1 (unreleased)
 ------------------
-
 - Calling context.register() and notify ObjectModifiedEvent after
   calling @sort/{field}, @delete/{field} and
   @delete/{field}/{file_key} to index changes
   [nilbacardit26]
-
+- Feat: Add metadata info to workflows
+- Fix: Update workflow vocabulary name
+- Feat: Update workflow vocabulary title attribute to use metadata
+  [rboixaderg]
 
 7.0.0 (2023-12-06)
 ------------------